+++ static/js/page_params.ts
@@ -25,6 +25,7 @@ export const page_params: {
     realm_enable_spectator_access: boolean;
     realm_invite_to_realm_policy: number;
     realm_invite_to_stream_policy: number;
+    realm_message_edit_history_visibility: number;
     realm_move_messages_between_streams_policy: number;
     realm_name_changes_disabled: boolean;
     realm_push_notifications_enabled: boolean;