+++ zerver/tests/test_events.py
@@ -2415,6 +2415,7 @@ class RealmPropertyActionTest(BaseAction):
             move_messages_between_streams_policy=Realm.COMMON_POLICY_TYPES,
             add_custom_emoji_policy=Realm.COMMON_POLICY_TYPES,
             delete_own_message_policy=Realm.COMMON_MESSAGE_POLICY_TYPES,
+            message_edit_history_visibility=Realm.MESSAGE_EDIT_HISTORY_VISIBILITY_TYPES,
         )
 
         vals = test_values.get(name)