+++ zerver/tests/test_message_dict.py
@@ -418,7 +418,7 @@ class MessageHydrationTest(ZulipTestCase):
             search_fields={},
             apply_markdown=True,
             client_gravatar=True,
-            allow_edit_history=False,
+            message_edit_history_visibility=Realm.MESSAGE_EDIT_HISTORY_VISIBILITY_NONE,
         )
 
         self.assert_length(messages, 2)
@@ -468,7 +468,7 @@ class MessageHydrationTest(ZulipTestCase):
             search_fields={},
             apply_markdown=True,
             client_gravatar=True,
-            allow_edit_history=False,
+            message_edit_history_visibility=Realm.MESSAGE_EDIT_HISTORY_VISIBILITY_NONE,
         )
         message = messages[0]
 
@@ -515,7 +515,7 @@ class TestMessageForIdsDisplayRecipientFetching(ZulipTestCase):
             search_fields={},
             apply_markdown=True,
             client_gravatar=True,
-            allow_edit_history=False,
+            message_edit_history_visibility=Realm.MESSAGE_EDIT_HISTORY_VISIBILITY_NONE,
         )
 
         self._verify_display_recipient(messages[0]["display_recipient"], [hamlet, cordelia])
@@ -536,7 +536,7 @@ class TestMessageForIdsDisplayRecipientFetching(ZulipTestCase):
             search_fields={},
             apply_markdown=True,
             client_gravatar=True,
-            allow_edit_history=False,
+            message_edit_history_visibility=Realm.MESSAGE_EDIT_HISTORY_VISIBILITY_NONE,
         )
 
         self._verify_display_recipient(
@@ -562,7 +562,7 @@ class TestMessageForIdsDisplayRecipientFetching(ZulipTestCase):
             search_fields={},
             apply_markdown=True,
             client_gravatar=True,
-            allow_edit_history=False,
+            message_edit_history_visibility=Realm.MESSAGE_EDIT_HISTORY_VISIBILITY_NONE,
         )
 
         self._verify_display_recipient(
@@ -596,7 +596,7 @@ class TestMessageForIdsDisplayRecipientFetching(ZulipTestCase):
             search_fields={},
             apply_markdown=True,
             client_gravatar=True,
-            allow_edit_history=False,
+            message_edit_history_visibility=Realm.MESSAGE_EDIT_HISTORY_VISIBILITY_NONE,
         )
 
         self._verify_display_recipient(